integrate with NalathniDragon's Appraisal mod
QudUX now automatically integrates with NalathniDragon's Appraisal mod if it detects that the Appraisal mod is installed. In this scenario, the ALT keybind for showing item values on the Inventory screen is only available after learning the Appraisal skill, and the Inventory screen calculates values using the custom formula from the Appraisal mod (instead of the usual integer rounding).